+++ app.js
@@ -2,7 +2,11 @@ import express from "express";
 import path, { delimiter } from "path";
 import { fileURLToPath } from "url";
 import danRoutes from "./routes/dan.js";
-import phuongQuanRoutes from "./routes/phuong.js";
+// phuong
+import diemDatQCPhuong from "./routes/phuong/diemDatQC-route.js";
+import bangQCPhuong from "./routes/phuong/bangQC-route.js";
+import baoCaoPhuong from "./routes/phuong/baoCao-route.js";
+
 import soQuanLyRoutes from "./routes/so/quanLy.js";
 import soHanhChinhRoutes from "./routes/so/hanhChinh.js";
 import soCanBoRoutes from "./routes/so/canbo.js";
@@ -76,7 +80,9 @@ canBoApp.get("/", (req, res) => {
 });
 
 canBoApp.use("/", authRouter);
-canBoApp.use("/", phuongQuanRoutes);
+canBoApp.use("/cac-diem-dat-quang-cao/", diemDatQCPhuong);
+canBoApp.use("/cac-bang-quang-cao/", bangQCPhuong);
+canBoApp.use("/cac-bao-cao/", baoCaoPhuong);
 canBoApp.use("/so/quanly", soQuanLyRoutes);
 canBoApp.use("/so/hanhchinh", soHanhChinhRoutes);
 canBoApp.use("/so/canbo", soCanBoRoutes);
