+++ views/phuong/reportDetails.ejs
@@ -114,7 +114,7 @@
   <div class="modal-dialog modal-lg">
     <form
       class="modal-content"
-      action="/update-report-status/<%=baoCaoId%>"
+      action="/cac-bao-cao/<%=baoCaoId%>/update-report-status"
       method="post"
       id="updateReport"
     >
